move tests for ModelReaders into cli/test_parse_*
