--- tests/cli/test_config_parser.py
@@ -1,63 +1,0 @@
-"""Test config load and parse
-"""
-import os
-from pytest import raises
-from smif.cli.parse_config import ConfigParser
-
-CONFIG_FIXTURES_DIR = os.path.join(os.path.dirname(__file__), "..", "fixtures", "config")
-
-def test_load_simple_config():
-    path = os.path.join(CONFIG_FIXTURES_DIR, "simple.yaml")
-    conf = ConfigParser(path)
-    assert conf.data["name"] == "test"
-
-def test_simple_validate_valid():
-    conf = ConfigParser()
-    conf.data = {"name": "test"}
-    conf.validate({
-        "type": "object",
-        "properties": {
-            "name": {
-                "type": "string"
-            }
-        },
-        "required": ["name"]
-    })
-
-def test_simple_validate_invalid():
-    conf = ConfigParser()
-    conf.data = {"name": "test"}
-
-    msg = "'nonexistent_key' is a required property"
-    with raises(ValueError, message=msg):
-        conf.validate({
-            "type": "object",
-            "properties": {
-                "nonexistent_key": {
-                    "type": "string"
-                }
-            },
-            "required": ["nonexistent_key"]
-        })
-
-def test_modelrun_config_validate():
-    path = os.path.join(CONFIG_FIXTURES_DIR, "modelrun_config.yaml")
-    conf = ConfigParser(path)
-
-    conf.validate_as_modelrun_config()
-
-def test_missing_timestep():
-    path = os.path.join(CONFIG_FIXTURES_DIR,
-                        "modelrun_config_missing_timestep.yaml")
-    conf = ConfigParser(path)
-
-    with raises(ValueError):
-        conf.validate_as_modelrun_config()
-
-def test_used_planning_needs_files():
-    path = os.path.join(CONFIG_FIXTURES_DIR,
-                        "modelrun_config_used_planning_needs_files.yaml")
-    conf = ConfigParser(path)
-
-    with raises(ValueError):
-        conf.validate_as_modelrun_config()
